flag_server/requirements_server.txt
python-dotenv==0.15.0
psycopg2==2.7.7
SQLAlchemy==1.3.20